pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = [
  "Name Surname <name.surname@mail.com>",
  "Name Surname <name.surname@mail.com>",
]
maintainers = [
  "Name Surname <name.surname@mail.com>",
  "Name Surname <name.surname@mail.com>",
]
license = "MIT"
readme = "README.md"
packages = [
  {include = "tools", from = "src"},
]
keywords = ["python", "template"]
classifiers = [  # Optional
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  # Pick your license as you wish
  "License :: OSI Approved :: MIT License",

  # Specify the Python versions you support here.
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]

[tool.poetry.dependencies]
python = "^3.11"

ipykernel="^6.28.0"
pathlib="^1.0.1"
GitPython="^3.1.40"
numpy="^1.26.3"
pandas="^2.1.4"
scipy="^1.11.4"
matplotlib="^3.8.2"
plotly="^5.18.0"
nbformat="^5.9.2"
pydantic="^2.5.3"

[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest="^7.4.4"
pytest-cov="^4.1.0"

[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
sphinx="^7.2.6"
sphinx-rtd-theme="^2.0.0"
myst-parser="^2.0.0"
numpydoc="^1.6.0"

[tool.poetry.group.deploy]
optional = true
[tool.poetry.group.deploy.dependencies]
fastapi="^0.108.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
